Fix typescript bindings for matchmaking messages.
Map<K, V> is actually an ES6 type and doesn't correspond to { K: V } at all. This was corrected so that this API can be used easily from within a TypeScript environment.
string_properties and numeric_properties were also made optional in matchmaker_add. I think this is correct as they are left off in an example in the documentation.